Support Docker & Portfolio Site
This commit adds two things:

1. Docker
There is a brand new "Dockerfile" which can be used to build an image
of this application.

2. Server (Express)
With the portfolio site changing over to Docker containers, this
application needs to be pluggable into the mix. This commit adds a
"server.js" file which runs a very basic Express server. Express was
therefore added to the "package.json" file.